Update draft-ietf-tcpm-rfc8312bis.md (#131)
* Update draft-ietf-tcpm-rfc8312bis.md

Take 2 at a PR is intended to amend the text to align with my understanding of the RFC-series. This normatively  would require ssthresh to be set from flight_size; however, this is explicitly relaxed in RFC7661, so I suggest wording should be careful around this topic. I retained the original text observation that in fact some cubic implementations are known to set ssthresh from cwnd.

* Update draft-ietf-tcpm-rfc8312bis.md

remove whitespace

Co-authored-by: Vidhi Goel <goel.vidhi07@gmail.com>

* Update draft-ietf-tcpm-rfc8312bis.md

That's OK

Co-authored-by: Vidhi Goel <goel.vidhi07@gmail.com>

* Update draft-ietf-tcpm-rfc8312bis.md

I think this may be obvious from the preceding section, but it seems technically more correct, so I'll commit this.

Co-authored-by: Vidhi Goel <goel.vidhi07@gmail.com>

* Update draft-ietf-tcpm-rfc8312bis.md

Now able to edit.  md file fixed to remove trailing whitespace.

* Add changelog

Co-authored-by: Vidhi Goel <goel.vidhi07@gmail.com>
Co-authored-by: Lars Eggert <lars@eggert.org>